+++ src/main/java/net/geforcemods/securitycraft/util/PlayerUtils.java
@@ -1,5 +1,8 @@
 package net.geforcemods.securitycraft.util;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
 
 import com.mojang.authlib.GameProfile;
@@ -176,5 +179,20 @@ public class PlayerUtils {
 		}
 
 		return null;
+	}
+
+	/**
+	 * Gets a list containing the player corresponding to an Owner
+	 *
+	 * @param owner The owner to get the player of
+	 * @return A list containing the player corresponding to the given owner, an empty list if no such player exists
+	 */
+	public static Collection<ServerPlayerEntity> getPlayerListFromOwner(Owner owner) {
+		ServerPlayerEntity player = getPlayerFromName(owner.getName());
+
+		if (player != null)
+			return Arrays.asList(player);
+
+		return new ArrayList<>();
 	}
 }