+++ Changelog.md
@@ -1,6 +1,7 @@
 --------------------------Changelog for v1.9.121 of SecurityCraft--------------------------
 
 - New: Config setting to disable the ability of the Universal Block Reinforcer to un-/reinforce blocks that are placed in the world
+- New: Option for the Portable Radar to only send notification messages to the owner instead of all members of the team
 - API: New method IPasscodeProtected#setPasscodeInAdjacentBlock for updating an adjacent block (e.g. the second half of a chest) with the passcode when setting it
 - Fix: Placing a Passcode-protected Chest or Barrel leads to unnecessary data being saved
 - Fix: Bouncing betties are rendered incorrectly